Made this fork to compile all the fixes/updates for the model importer. More to come shortly.
Updated setup.py to target the specific pyglm version that works with the tool. Thanks to Kamen Rider.
Updated the Material Database file the importer uses to allow for proper imports of stage materials thanks to neociano.
Updated metadata.py to use an updated command instead of the depreciated one.