Refactored Joint MT Attributes to read from Pose Bone. Also adjusted Materials so nodes aren't all top of one another.
Attempted to reverse the bone related operations to reverse engineer the importing process. to do this:
Had to refactor the MT attributes to use Pose Bones to store and read from to so
the parent MT attributes can be accessed for code later.
Adjusted Materials so nodes aren't all top of one another.